requirements-dashboard.txt
streamlit
kafka-python
cassandra-driver
pandas